fix undefined integer shift
From: Stefan Bühler <stbuehler@web.de>

git-svn-id: svn://svn.lighttpd.net/lighttpd/branches/lighttpd-1.4.x@3011 152afb58-edef-0310-8abb-c4023f1b3aa9

+++ src/stat_cache.c
@@ -249,7 +249,7 @@ static uint32_t hashme(buffer *str) {
 		hash = ((hash << 5) + hash) + *s;
 	}
 
-	hash &= ~(1 << 31); /* strip the highest bit */
+	hash &= ~(((uint32_t)1) << 31); /* strip the highest bit */
 
 	return hash;
 }
